Changed Architecture as:  - PubnativeAPIV3ResponseModel contains list of PubnativeAPIV3AdModel  - PubnativeAdModel contains a PubnativeAdDataModel  - PubnativeAPIV3AdModel implements PubnativeAdDataModel  - PubnativeAdDataModel only contains abstract getters  - PubnativeAPIV3AdModel contains the fields of the real data
