+++ library/src/main/java/net/pubnative/library/request/model/PubnativeAPIV3AdModel.java
@@ -1,56 +1,169 @@
 package net.pubnative.library.request.model;
 
+import android.text.TextUtils;
+import android.util.Log;
+
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class PubnativeAPIV3AdModel extends PubnativeAdDataModel {
+public class PubnativeAPIV3AdModel implements PubnativeAdDataModel {
 
-    public interface AdData {
-        String getClickUrl();
+    private static String                   TAG     = PubnativeAPIV3AdModel.class.getSimpleName();
 
-        String getTitle();
+    protected String                        link;
+    protected List<PubnativeAPIV3DataModel> assets;
+    protected List<PubnativeAPIV3DataModel> beacons;
+    protected List<PubnativeAPIV3DataModel> meta;
 
-        String getDescription();
+    public String getClickUrl() {
 
-        String getCta();
-
-        String getRating();
-
-        PubnativeImage getIcon();
-
-        PubnativeImage getBanner();
-
-        HashMap<String, String> getMetaField(String metaType);
-
-        List<String> getAllBeacons();
+        Log.v(TAG, "getClickUrl");
+        return link;
     }
 
-    public interface AdBeacon {
-        List<String> getAllBeacons();
-        List<String> getJsBeacons();
+    public String getTitle() {
+
+        Log.v(TAG, "getTitle");
+        String title = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.TITLE);
+        if(asset != null) {
+            title = asset.data.get("text");
+        }
+        return title;
     }
 
-    /**
-     * These are the various types of assets
-     */
-    public interface AssetType {
+    public String getDescription() {
 
-        String TITLE        = "title";
-        String DESCRIPTION  = "description";
-        String CTA          = "cta";
-        String RATING       = "rating";
-        String ICON         = "icon";
-        String BANNER       = "banner";
+        Log.v(TAG, "getDescription");
+        String description = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.DESCRIPTION);
+        if(asset != null) {
+            description = asset.data.get("text");
+        }
+        return description;
     }
 
-    /**
-     * These are meta fields
-     */
-    public interface MetaType {
+    public String getCta() {
 
-        String POINTS       = "points";
-        String REVENU_MODEL = "revenuemodel";
-        String CAMPAIGN_ID  = "campaignid";
-        String CREATIVE_ID  = "creativeid";
+        Log.v(TAG, "getCta");
+        String cta = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.CTA);
+        if(asset != null) {
+            cta = asset.data.get("text");
+        }
+        return cta;
+    }
+
+    public String getRating() {
+
+        Log.v(TAG, "getRating");
+        String rating = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.RATING);
+        if(asset != null) {
+            rating = asset.data.get("number");
+        }
+        return rating;
+    }
+
+    public PubnativeImage getIcon() {
+
+        Log.v(TAG, "getIcon");
+        PubnativeImage icon = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.ICON);
+        if(asset != null) {
+
+            icon = new PubnativeImage();
+            icon.width = Integer.valueOf(asset.data.get("w"));
+            icon.height = Integer.valueOf(asset.data.get("h"));
+            icon.url = asset.data.get("url");
+        }
+        return icon;
+    }
+
+    public PubnativeImage getBanner() {
+
+        Log.v(TAG, "getBanner");
+        PubnativeImage banner = null;
+        PubnativeAPIV3DataModel asset = findAsset(PubnativeAPIV3AdModel.AssetType.BANNER);
+        if(asset != null) {
+
+            banner = new PubnativeImage();
+            banner.width = Integer.valueOf(asset.data.get("w"));
+            banner.height = Integer.valueOf(asset.data.get("h"));
+            banner.url = asset.data.get("url");
+        }
+        return banner;
+    }
+
+    public HashMap<String, String> getMetaField(String metaType) {
+
+        Log.v(TAG, "getMetaField");
+        HashMap<String, String> result = null;
+
+        if(meta != null) {
+            for(PubnativeAPIV3DataModel model: meta) {
+                if(model.type.equals(metaType)) {
+                    result = model.data;
+                }
+            }
+        }
+        return result;
+    }
+
+    public List<String> getAllBeacons() {
+
+        Log.v(TAG, "getAllBeacons");
+        List<String> allBeacons = null;
+        if (beacons != null) {
+            allBeacons = new ArrayList<String>();
+            for (PubnativeAPIV3DataModel beacon : beacons) {
+                if(!TextUtils.isEmpty(beacon.data.get("url"))) {
+                    allBeacons.add(beacon.data.get("url"));
+                }
+            }
+        }
+        if(assets != null) {
+            if(allBeacons == null) {
+                allBeacons = new ArrayList<String>();
+            }
+            for(PubnativeAPIV3DataModel model: assets) {
+                if(model.data.containsKey("tracking") && !TextUtils.isEmpty(model.data.get("tracking"))) {
+                    allBeacons.add(model.data.get("tracking"));
+                }
+            }
+        }
+        return allBeacons;
+    }
+
+    public List<String> getJsBeacons() {
+        Log.v(TAG, "getJsBeacons");
+        List<String> jsBeacons = null;
+        if (beacons != null) {
+            jsBeacons = new ArrayList<String>();
+            for (PubnativeAPIV3DataModel beacon : beacons) {
+                if(!TextUtils.isEmpty(beacon.data.get("js"))) {
+                    jsBeacons.add(beacon.data.get("js"));
+                }
+            }
+        }
+        return jsBeacons;
+    }
+
+    //==============================================================================================
+    // Private
+    //==============================================================================================
+    private PubnativeAPIV3DataModel findAsset(String assetType) {
+
+        PubnativeAPIV3DataModel result = null;
+
+        if(assets != null) {
+            for(PubnativeAPIV3DataModel model: assets) {
+                if(model.type.equals(assetType)) {
+                    result = model;
+                }
+            }
+        }
+        return result;
     }
 }
