+++ library/src/main/java/net/pubnative/library/request/model/PubnativeAPIV3DataModel.java
@@ -1,0 +1,9 @@
+package net.pubnative.library.request.model;
+
+import java.util.HashMap;
+
+public class PubnativeAPIV3DataModel {
+
+    protected String                  type;
+    protected HashMap<String, String> data;
+}
