+++ library/src/main/java/net/pubnative/library/request/model/PubnativeAdDataModel.java
@@ -1,8 +1,45 @@
 package net.pubnative.library.request.model;
 
 import java.util.HashMap;
+import java.util.List;
 
-public abstract class PubnativeAdDataModel {
-    protected String                    type;
-    protected HashMap<String, String>   data;
+public interface PubnativeAdDataModel {
+
+    String getClickUrl();
+
+    String getTitle();
+
+    String getDescription();
+
+    String getCta();
+
+    String getRating();
+
+    PubnativeImage getIcon();
+
+    PubnativeImage getBanner();
+
+    HashMap<String, String> getMetaField(String metaType);
+
+    List<String> getAllBeacons();
+
+    List<String> getJsBeacons();
+
+    interface AssetType {
+
+        String TITLE        = "title";
+        String DESCRIPTION  = "description";
+        String CTA          = "cta";
+        String RATING       = "rating";
+        String ICON         = "icon";
+        String BANNER       = "banner";
+    }
+
+    interface MetaType {
+
+        String POINTS       = "points";
+        String REVENU_MODEL = "revenuemodel";
+        String CAMPAIGN_ID  = "campaignid";
+        String CREATIVE_ID  = "creativeid";
+    }
 }
