--- library/src/main/java/net/pubnative/library/request/model/PubnativeExtDataModel.java
@@ -1,4 +1,0 @@
-package net.pubnative.library.request.model;
-
-public abstract class PubnativeExtDataModel {
-}
